backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
torch>=2.9.0
torchvision>=0.24.0
Pillow>=10.3.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0